+++ tests/builtin_test_functions/test_sobol_g.py
@@ -72,7 +72,7 @@ def test_compute_variance(spatial_dimension, params_selection):
     var_mc = np.var(yy)
 
     # Analytical variance
-    params = my_fun.parameters
+    params = my_fun.parameters["aa"]
     var_ref = np.prod((4 / 3 + 2 * params + params**2) / (1 + params) ** 2) - 1
 
     # Assertion (no need to be very ambitious with the tolerance)